Simplify ActiveConfiguredProjectsLoader
This class monitors CPS's `ActiveConfigurationGroupSource` block and forces CPS to eagerly load the project for all configurations in the project.

CPS is looking at moving this logic to within CPS itself, giving greater control over when configured projects (across all dimensions) are loaded, for example to prevent various race conditions.

The previous implementation had two distinct behaviours:

1. Use `[ProjectAutoLoad]` to link an action to the `ActiveConfigurationGroupSource` that forces loads to occur.
2. Export a `ChainedProjectValueDataSourceBase<>`, which allows it to participate in Dataflow so that other blocks can subscribe to its output (`IEnumerable<ConfiguredProject>`).

This change keeps 1 and removes 2. That way, when CPS adds the behaviour of 1, we can be sure we don't need the behaviour of 2.