Alias ImportCardinality via global implicit using
MEF types exist in two namespaces, and we generally use only one of them. However we do use the `ImportCardinality` enum from the other. Adding an import alias for that type means we can avoid adding a `using` directive for the entire namespace we want to avoid, reducing the chance of accidentally using the wrong kind of MEF attributes.