Remove explicit contract name from MEF part
Contract names are used when the CLR type is not enough to uniquely identify the part for import. For example, when we use an interface in multiple scopes, we use explicit contract names to control which exports of the interface to import in each scope.

In the case of `IDependenciesRuleHandler`, there is no such need for this contract name. This commit removes it.