Write file reference with correct capitalisation
When dragging a file from Solution Explorer into the Managed Resources editor, the file path is written to the `.resx` file in lower case.

This is because the "drop data" provided by Solution Explorer contains a lower-case path (both for SDK-style and legacy projects).

To fix this, we walk the path, using `Path.EnumerateFileSystemEntries` to produce a correctly cased path for the file being dragged. Unfortunately there's no better way I could find to achieve this. I tried quite a few different approaches and this seemed like the least rubbish. The approach used in `GetFileNameInActualCase` elsewhere in the resource editor only handles the file name, not any directories above it.