Annotate MEF part as non-shared
The language service's CompileItemHandler is created multiple times within the unconfigured project -- once per project configuration slice. We annotate it here with an attribute to advertise this fact.

This shouldn't have any impact to the current code as we use an ExportFactory to create instances of this part as needed, however annotating it is good practice and makes the type consistent with its peers.