Fix FUTDC solution cache population for DevBox
When running `devenv <sln> /populateSolutionCache /build`, the FUTDC should produce its cache file such that the first build using the generated caches is considered up-to-date.

The FUTDC has historically disabled itself when running in command line mode. Historically command line mode has not run any design-time builds, and therefore no FUTDC snapshot data becomes available. The FUTDC would avoid waiting for snapshots when it detected command line mode is active.

Solution cache population is now an exception to this. It's a form of command line mode for where design-time builds do run, and for which we do want the FUTDC to run in full.

This change teaches the `UpToDateCheckHost.HasDesignTimeBuildsAsync` method about the `/populateSolutionCache` command line argument, so that the `.futdcache.v2` file is correctly populated in this mode.