Fix early signalling of loaded state
Our `UnconfiguredProjectTasksService` class listens for project factory completion and signals this to other components that may wait on that status.

The class uses `[ProjectAutoLoad]` to receive this notification from CPS, however it was asking to be notified _before_ the project factory completed rather than _after_.

One identified consequence of this is that the `SDKVersionTelemetryServiceComponent` class would wake up earlier than intended, request `ActiveConfiguredProjectProperties`, then immediately block. This makes a thread-pool thread unavailable for other productive work during project/solution load.

The fix here is to modify the `[ProjectAutoLoad]` registration so we are notified _after_ the project factory completes.

I've reviewed components impacted by this change and believe the change to be safe.