Extract code for launch settings
When launching a project using a launch profile we use information from the project itself--the path to the .exe to run, the command line arguments, and the working directory--and then supplement or replace them with information from the profile. I do not want to create a copy of this code in our launch profiles support for VS Code as that may lead the implementations to diverge over time, which in turn may impact the user's ability to use the same launchSettings.json in both VS for Windows and VS Code.

Instead, this change extracts out some of the parts that should definitely remain consistent between VSW and VS Code, and pulls them down to the Microsoft.VisualStudio.ProjectSystem.Managed layer where they can be used by both. I've also added doc comments explaining the behavior of the methods.

I tried to make this refactoring as mechanical as possible by converting any dependencies on fields to method parameters, then making the method static, and then moving it to `ProjectAndExectuableLaunchHandlerHelpers`. I've tried to avoid any significant changes in the logic, even where it currently seems weird or undesirable. I've also added a small number of unit tests, though many scenarios are also covered by the existing `ProjectLaunchTargetsProviderTests`.