Remove OutputTypeChecker use
Remove the explicit creation and use of the `OutputTypeChecker` class from `ProjectLaunchTargetsProviderTests` and replace it with a mock of `IOutputTypeChecker`. The main benefit here is that it makes it clear that the `ProjectLaunchTargetsProvider` doesn't care about the "OutputType" property per se, but really only cares if the project produces a traditional library, a console app, or "something else", as those are the only possibilities that are actually exposed by `IOutputTypeChecker`.

This does raise the question about whether the behavior is really what we want in the case of "something else". E.g., do we really want to have different behavior for a "library" project versus a "winmdobj"? I don't attempt to answer that question here, but at least the tests more clearly document the existing behavior.