Remove dependency on the project tree
Fixes AB#1815584.

The `LaunchSettingsProvider` looks for the launchSettings.json file within the project's "app designer" folder, and so to read the file it needs to find the location of that folder on disk. Ultimately this is controlled by the "AppDesignerFolder" property in MSBuild. In the case of C# this defaults to "Properties"; for VB it is called "My Project".

Currently the `LaunchSettingsProvider` does not read this property directly. Instead, it imports the `IAppDesignerFolderSpecialFileProvider`, and the implementation of that (`AppDesignerFolderSpecialFileProvider`) utilizes project tree. It searches the project tree for certain flags denoting the app designer folder; these flags were previously set by the `AppDesignerFolderProjectTreePropertiesProvider`. It is that type that makes use of the underlying MSBuild property values.

These extra levels of abstraction add cost and complexity without, as best I can tell, buying us much in the way of functionality. As it stands we can't determine the path to the launchSettings.json until the project tree has been created by CPS and updated by the `AppDesignerFolderProjectTreePropertiesProvider`, and in some cases we may force the creation of the project tree where it is otherwise not needed.

In this change we cut out the middle layers and go directly to the MSBuild properties via the `ProjectProperties` type.

Note that this does not correct other, long-standing issues with this code. Notably we will retrieve the path once and cache it, and as such will not react to any project changes that might change the location of the app designer folder. To do that would require significant rework to acquire the app designer folder path through a dataflow, which would also require that changes to how we manage the file watcher on the launchSettings.json. That work is already tracked by #2316.