Add a platform check
Related to #1824245.

The `ProjectAndExecutableLaunchHandlerHelpers` currently ensure that when
we have the name of an executable to run it ends in ".exe". This makes
sense on Windows, but not on Linux or Mac where executable files typically
do not end in .exe. Here we add a platform check and only append ".exe"
on Windows.