Fix FUTD check of copy items when building for debug
When the user presses F5 the Solution Build Manager (SBM) calls the Fast Up-to-date Check (FUTDC) twice, for some reason I don't fully understand. If you invoke build directly, it's only called once.

In the debug case, the first of the two calls occurs *before* any SBM events are fired, which breaks an assumption that was made about the sequence of events. Specifically we assumed that the FUTDC would only be called between calls to `IVsUpdateSolutionEvents.UpdateSolution_StartUpdate` and `IVsUpdateSolutionEvents.UpdateSolution_Done`. But that's not the case when the user presses F5.

It turns out that this surfaces an issue only for `CopyToOutputDirectory` items, as its only for these items that we use the shared `SolutionBuildContext` cache of timestamps. That cache helps considerably with the performance of Build Acceleration across the solution, where the same files are checked repeatedly within the solution.

This change fixes the issue by creating a fake "solution build starting" event when we detect that no such event occurred prior to the FUTDC being called.