Avoid flickering diagnostic levels
When evaluation-only data is available for a dependency item, it's unlikely that we have `DiagnosticLevel` metadata. In such cases, we would previously conclude a level of `None`. However if the prior build data had a level of `Warning` and the subsequent build data will have a level of `Warning`, then the icon will flicker from warning to no warning back to warning again.

This avoids this flickering by maintaining the diagnostic level that existed previously when evaluation data does not include an explicit override.