Fix CodeQL in build
The CodeQL task in auto-injected into our official builds in order to check for known code vulnerabilities. A bug in CodeQL was causing "dotnet test" to fail (see https://twcsecurityassurance.visualstudio.com/Semmle/_workitems/edit/24228), and while it has been fixed in a newer version of CodeQL it will be a while before the pipeline task picks up the new version. In the meantime we can specify a version with the fix by setting Codeql.CLIVersion and Codeql.CLIHash in our official.yml file.

However, the injected CodeQL task is now complaining that "Hash verification failed" and disabling CodeQL. The issue is the text "// sha256 hash for windows CLI" at the end of the CLIHash value; it is being treated as part of the hash rather than a comment. This change removes the comment.