Don't hang when update handling throws, and report NFE
Components such as CodeModel can wait on the language service to initialise. In particular, they need the first "primary" workspace to be set, so that it can be interacted with.

Currently that state is signalled with a `TaskCompletionSource`.

If the first data update causes an exception in the handler, the previous code would neither complete nor fault that `TaskCompletionSource`, leaving any waiting code to wait indefinitely (hang).

This change ensures that any fault on the handling block causes that `TaskCompletionSource` to be faulted, if it isn't already complete.

We also register the failure with the project fault handler, so that an NFE is sent via telemetry, and a gold bar is displayed to the user indicating the the project is in a state where it will have limited functionality.