Update conditions to show Native AOT properties.
I have changed the project capability condition to include _NativeAOTProperties_ to projects targeting .NET 8 or higher. We later exclude/include class libraries when checking the visibility condition:
The _IsTrimmable_ and _IsAotCompatible_ properties should only be shown to class library projects.
The _Trimming_ and _PublishAot_ properties should not be shown to class library projects.