Avoid NFE when ConfigurationGeneral rule is modified to remove expected property
The project system uses rule files to define a schema for the properties and items it needs to extract from the project. We use these rules in code when retrieving such values.

The rules are also extension points, and project types may replace or extend these rules for their own purposes.

Replacing a core rule file (like `ConfigurationGeneral`) is a bad idea as it can fall out of date when we take dependencies upon properties that are expected to be there, and the replacement rule does not include them.

That appears to be what's happening from our NFE data. Some project configuration is using the FUTDC while not defining the `TargetPath` property in the `ConfigurationGeneral` rule.

I'll investigate this in more depth separately as it seems like another component might have an issue that could resurface elsewhere. In the meantime, this change makes the FUTDC tolerant of this missing value. The consequence of it being absent is the checking of copy items won't be as extensive, and build acceleration would be disabled for that project and for any that reference it.