Fix empty dependency group node for external dependencies
Implementations of the `IProjectDependenciesSubTreeProvider` interface may extend the dependencies tree with additional dependency types.

The rewrite in 17.7 introduced a subtle change in bevahiour where removing the last dependency in a group added via this interface would leave the group node visible in the tree.

This change ensures we remove the group altogether when the last item is removed.