Improve installed runtime detection
This fixes an issue where the `RegistryKey.OpenSubKey` can return `null` in case of failure, which would then cause a fault. Instead we now have the method return `null` when installed runtimes could not be found, and the code that consumes that value has some fallback logic for the `null` case. It attempts to detect when the component is likely a runtime, and errs on the side of not reporting a missing runtime, so we don't bother users to install something they already have.

This change also fixes incorrect parsing of runtime versions that existed previously, where any version string that contained prerelease information (such as `8.0.0-preview.7.23375.6`) would fail to be detected correctly.