MEF component will never be null
Previously the FUTDC was split between two layers, with `IUpToDateCheckStatePersistence` coming from the VS layer. The consumer didn't assume that the VS layer was present, so used `AllowDefault = true` on that import.

Since then, we moved the FUTDC into the VS layer. So now, we can assume `IUpToDateCheckStatePersistence` will always be present and non-null.